requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
